+++ tests/Unit/CMBoundaryGroups.tests.ps1
@@ -211,9 +211,77 @@ try
             Ensure        = 'Absent'
         }
 
-        Describe "$moduleResourceName\Get-TargetResource" {
-            Mock -CommandName Import-ConfigMgrPowerShellModule
-            Mock -CommandName Set-Location
+        Describe 'ConfigMgrCBDsc - DSC_CMBoundaryGroups\Get-TargetResource' -Tag 'Get' {
+            BeforeAll {
+                $getInput = @{
+                    SiteCode      = 'Lab'
+                    BoundaryGroup = 'TestGroup'
+                }
+
+                $getBoundaryGroupOutput = @{
+                    Name    = 'TestGroup'
+                    GroupID = 16777229
+                }
+
+                $boundarySiteSystem = @(
+                    @{
+                        GroupID       = 16777229
+                        ServerNalPath = '["Display=\\DP01.contoso.com\"]MSWNET:["SMS_SITE=LAB"]\\DP01.contoso.com\'
+                    }
+                    @{
+                        GroupID       = 16777229
+                        ServerNalPath = '["Display=\\DP02.contoso.com\"]MSWNET:["SMS_SITE=LAB"]\\DP02.contoso.com\'
+                    }
+                )
+
+                $getBoundaryOutput = @(
+                    @{
+                        BoundaryId   = 16777231
+                        BoundaryType = 3
+                        Value        = '10.1.1.1-10.1.1.255'
+                    }
+                    @{
+                        BoundaryId   = 16777232
+                        BoundaryType = 0
+                        Value        = '10.1.3.0'
+                    }
+                    @{
+                        BoundaryId   = 16777233
+                        BoundaryType = 2
+                        Value        = 'First-Site'
+                    }
+                )
+
+                $mockBoundaryMembers = @(
+                    (New-CimInstance -ClassName DSC_CMBoundaryGroupsBoundaries `
+                        -Namespace root/microsoft/Windows/DesiredStateConfiguration `
+                        -Property @{
+                            'Value' = '10.1.1.1-10.1.1.255'
+                            'Type'  = 'IPRange'
+                        } `
+                        -ClientOnly
+                    ),
+                    (New-CimInstance -ClassName DSC_CMCollectionQueryRules `
+                        -Namespace root/microsoft/Windows/DesiredStateConfiguration `
+                        -Property @{
+                            'Value' = '10.1.3.0'
+                            'Type'  = 'IPSubnet'
+                        } `
+                        -ClientOnly
+                    ),
+                    (New-CimInstance -ClassName DSC_CMCollectionQueryRules `
+                        -Namespace root/microsoft/Windows/DesiredStateConfiguration `
+                        -Property @{
+                            'Value' = 'First-Site'
+                            'Type'  = 'ADSite'
+                        } `
+                        -ClientOnly
+                    )
+                )
+
+                Mock -CommandName Import-ConfigMgrPowerShellModule
+                Mock -CommandName Set-Location
+            }
 
             Context 'When retrieving boundary group settings' {
 
@@ -221,12 +289,14 @@ try
                     Mock -CommandName Get-CMBoundaryGroup -MockWith { $null }
                     Mock -CommandName Get-CMBoundary -MockWith { $null }
                     Mock -CommandName ConvertTo-CimBoundaries -MockWith { $null }
+                    Mock -CommandName Get-CMBoundaryGroupSiteSystem -MockWith { $null }
 
                     $result = Get-TargetResource @getInput
                     $result               | Should -BeOfType System.Collections.HashTable
                     $result.SiteCode      | Should -Be -ExpectedValue 'Lab'
                     $result.BoundaryGroup | Should -Be -ExpectedValue 'TestGroup'
                     $result.Boundaries    | Should -Be -ExpectedValue $null
+                    $result.SiteSystems   | Should -Be -ExpectedValue $null
                     $result.Ensure        | Should -Be -ExpectedValue 'Absent'
                 }
 
@@ -234,12 +304,14 @@ try
                     Mock -CommandName Get-CMBoundaryGroup -MockWith { $getBoundaryGroupOutput }
                     Mock -CommandName Get-CMBoundary -MockWith { $null }
                     Mock -CommandName ConvertTo-CimBoundaries -MockWith { $null }
+                    Mock -CommandName Get-CMBoundaryGroupSiteSystem -MockWith { $null }
 
                     $result = Get-TargetResource @getInput
                     $result               | Should -BeOfType System.Collections.HashTable
                     $result.SiteCode      | Should -Be -ExpectedValue 'Lab'
                     $result.BoundaryGroup | Should -Be -ExpectedValue 'TestGroup'
                     $result.Boundaries    | Should -Be -ExpectedValue $null
+                    $result.SiteSystems   | Should -Be -ExpectedValue $null
                     $result.Ensure        | Should -Be -ExpectedValue 'Present'
                 }
 
@@ -247,6 +319,7 @@ try
                     Mock -CommandName Get-CMBoundaryGroup -MockWith { $getBoundaryGroupOutput }
                     Mock -CommandName Get-CMBoundary -MockWith { $getBoundaryOutput }
                     Mock -CommandName ConvertTo-CimBoundaries -MockWith { $mockBoundaryMembers }
+                    Mock -CommandName Get-CMBoundaryGroupSiteSystem -MockWith { $boundarySiteSystem }
 
                     $result = Get-TargetResource @getInput
                     $result                  | Should -BeOfType System.Collections.HashTable
@@ -254,20 +327,178 @@ try
                     $result.BoundaryGroup    | Should -Be -ExpectedValue 'TestGroup'
                     $result.Boundaries       | Should -BeOfType '[Microsoft.Management.Infrastructure.CimInstance]'
                     $result.Boundaries.Count | Should -Be -ExpectedValue 3
+                    $result.SiteSystems      | Should -Be -ExpectedValue @('DP01.contoso.com','DP02.contoso.com')
                     $result.Ensure           | Should -Be -ExpectedValue 'Present'
                 }
             }
         }
 
-        Describe "$moduleResourceName\Set-TargetResource" {
-            Context 'When Set-TargetResource runs successfully' {
+        Describe 'ConfigMgrCBDsc - DSC_CMBoundaryGroups\Set-TargetResource' -Tag 'Set' {
+            BeforeAll {
+                $mockBoundaryMembers = @(
+                    (New-CimInstance -ClassName DSC_CMBoundaryGroupsBoundaries `
+                        -Namespace root/microsoft/Windows/DesiredStateConfiguration `
+                        -Property @{
+                            'Value' = '10.1.1.1-10.1.1.255'
+                            'Type'  = 'IPRange'
+                        } `
+                        -ClientOnly
+                    ),
+                    (New-CimInstance -ClassName DSC_CMCollectionQueryRules `
+                        -Namespace root/microsoft/Windows/DesiredStateConfiguration `
+                        -Property @{
+                            'Value' = '10.1.3.0'
+                            'Type'  = 'IPSubnet'
+                        } `
+                        -ClientOnly
+                    ),
+                    (New-CimInstance -ClassName DSC_CMCollectionQueryRules `
+                        -Namespace root/microsoft/Windows/DesiredStateConfiguration `
+                        -Property @{
+                            'Value' = 'First-Site'
+                            'Type'  = 'ADSite'
+                        } `
+                        -ClientOnly
+                    )
+                )
+
+                $mockInputBoundaryRange = @(
+                    (New-CimInstance -ClassName DSC_CMBoundaryGroupsBoundaries `
+                        -Namespace root/microsoft/Windows/DesiredStateConfiguration `
+                        -Property @{
+                            'Value' = '10.1.1.1-10.1.1.255'
+                            'Type'  = 'IPRange'
+                        } `
+                        -ClientOnly
+                    )
+                )
+
+                $mockOutputBoundarySubnet = @(
+                    (New-CimInstance -ClassName DSC_CMBoundaryGroupsBoundaries `
+                        -Namespace root/microsoft/Windows/DesiredStateConfiguration `
+                        -Property @{
+                            'Value' = '10.1.3.0'
+                            'Type'  = 'IPSubnet'
+                        } `
+                        -ClientOnly
+                    )
+                )
+
+                $mockInputBoundarySubnet = @(
+                    (New-CimInstance -ClassName DSC_CMBoundaryGroupsBoundaries `
+                        -Namespace root/microsoft/Windows/DesiredStateConfiguration `
+                        -Property @{
+                            'Value' = '10.1.3.1/24'
+                            'Type'  = 'IPSubnet'
+                        } `
+                        -ClientOnly
+                    )
+                )
+
+                $mockInputBoundaryADSiteAdd = @(
+                    (New-CimInstance -ClassName DSC_CMBoundaryGroupsBoundaries `
+                        -Namespace root/microsoft/Windows/DesiredStateConfiguration `
+                        -Property @{
+                            'Value' = 'Second-Site'
+                            'Type'  = 'ADSite'
+                        } `
+                        -ClientOnly
+                    )
+                )
+
+                $mockInputBoundaryADSite = @(
+                    (New-CimInstance -ClassName DSC_CMBoundaryGroupsBoundaries `
+                        -Namespace root/microsoft/Windows/DesiredStateConfiguration `
+                        -Property @{
+                            'Value' = 'First-Site'
+                            'Type'  = 'ADSite'
+                        } `
+                        -ClientOnly
+                    )
+                )
+
+                $getReturnAbsent = @{
+                    SiteCode      = 'Lab'
+                    BoundaryGroup = 'TestGroup'
+                    Boundaries    = $null
+                    SiteSystens   = $null
+                    Ensure        = 'Absent'
+                }
+
+                $setTestInputAdd = @{
+                    SiteCode       = 'Lab'
+                    BoundaryGroup  = 'TestGroup'
+                    Boundaries     = $mockInputBoundarySubnet
+                    BoundaryAction = 'Add'
+                }
+
+                $getReturn = @{
+                    SiteCode      = 'Lab'
+                    BoundaryGroup = 'TestGroup'
+                    Boundaries    = $mockBoundaryMembers
+                    SiteSystems   = @('DP01.contoso.com', 'DP02.contoso.com')
+                    Ensure        = 'Present'
+                }
+
+                $setTestInputAddValue = @{
+                    SiteCode       = 'Lab'
+                    BoundaryGroup  = 'TestGroup'
+                    Boundaries     = $mockInputBoundaryADSiteAdd
+                    BoundaryAction = 'Add'
+                }
+
+                $setTestInputMatch = @{
+                    SiteCode       = 'Lab'
+                    BoundaryGroup  = 'TestGroup'
+                    Boundaries     = $mockInputBoundaryRange
+                    BoundaryAction = 'Match'
+                }
+
+                $setTestInputRemove = @{
+                    SiteCode       = 'Lab'
+                    BoundaryGroup  = 'TestGroup'
+                    Boundaries     = $mockInputBoundarySubnet
+                    BoundaryAction = 'Remove'
+                }
+
+                $setTestInputGroupOnlyAbsent = @{
+                    SiteCode      = 'Lab'
+                    BoundaryGroup = 'TestGroup'
+                    Ensure        = 'Absent'
+                }
+
+                $matchSiteSystems = @{
+                    SiteCode             = 'Lab'
+                    BoundaryGroup        = 'TestGroup'
+                    SiteSystems          = 'DP01.Contoso.com','DP03.contoso.com'
+                }
+
+                $includeSiteSystems = @{
+                    SiteCode             = 'Lab'
+                    BoundaryGroup        = 'TestGroup'
+                    SiteSystemsToInclude = 'DP01.Contoso.com','DP03.contoso.com'
+                }
+
+                $siteSystemsIncludeExclude = @{
+                    SiteCode             = 'Lab'
+                    BoundaryGroup        = 'TestGroup'
+                    SiteSystemsToInclude = 'DP02.contoso.com'
+                    SiteSystemsToExclude = 'DP02.contoso.com'
+                }
+
+                $siteInEx = 'SiteSystemsToInclude and SiteSystemsToExclude contain the same member DP02.contoso.com.'
+
                 Mock -CommandName Import-ConfigMgrPowerShellModule
                 Mock -CommandName Set-Location
                 Mock -CommandName New-CMBoundaryGroup
                 Mock -CommandName Add-CMBoundaryToGroup
                 Mock -CommandName Remove-CMBoundaryGroup
                 Mock -CommandName Remove-CMBoundaryFromGroup
+                Mock -CommandName Set-CMBoundaryGroup
+                Mock -CommandName Get-CMSiteSystemServer -MockWith { $true }
+            }
 
+            Context 'When Set-TargetResource runs successfully' {
                 It 'Should call expected commands for new boundary group and adding boundary' {
                     Mock -CommandName Get-TargetResource -MockWith { $getReturnAbsent }
                     Mock -CommandName Convert-BoundariesIPSubnets -MockWith { $mockOutputBoundarySubnet }
@@ -283,6 +514,8 @@ try
                     Assert-MockCalled Get-BoundaryInfo -Exactly -Times 1 -Scope It
                     Assert-MockCalled Remove-CMBoundaryGroup -Exactly -Times 0 -Scope It
                     Assert-MockCalled Remove-CMBoundaryFromGroup -Exactly -Times 0 -Scope It
+                    Assert-MockCalled Get-CMSiteSystemServer -Exactly -Times 0 -Scope It
+                    Assert-MockCalled Set-CMBoundaryGroup -Exactly -Times 0 -Scope It
                 }
 
                 It 'Should call expected commands for adding a new boundary to the group' {
@@ -300,6 +533,8 @@ try
                     Assert-MockCalled Get-BoundaryInfo -Exactly -Times 1 -Scope It
                     Assert-MockCalled Remove-CMBoundaryGroup -Exactly -Times 0 -Scope It
                     Assert-MockCalled Remove-CMBoundaryFromGroup -Exactly -Times 0 -Scope It
+                    Assert-MockCalled Get-CMSiteSystemServer -Exactly -Times 0 -Scope It
+                    Assert-MockCalled Set-CMBoundaryGroup -Exactly -Times 0 -Scope It
                 }
 
                 It 'Should call expected commands for boundary set to match removing two additional boundaries' {
@@ -317,6 +552,8 @@ try
                     Assert-MockCalled Get-BoundaryInfo -Exactly -Times 2 -Scope It
                     Assert-MockCalled Remove-CMBoundaryGroup -Exactly -Times 0 -Scope It
                     Assert-MockCalled Remove-CMBoundaryFromGroup -Exactly -Times 2 -Scope It
+                    Assert-MockCalled Get-CMSiteSystemServer -Exactly -Times 0 -Scope It
+                    Assert-MockCalled Set-CMBoundaryGroup -Exactly -Times 0 -Scope It
                 }
 
                 It 'Should call expected commands when removing a boundary' {
@@ -334,6 +571,8 @@ try
                     Assert-MockCalled Get-BoundaryInfo -Exactly -Times 1 -Scope It
                     Assert-MockCalled Remove-CMBoundaryGroup -Exactly -Times 0 -Scope It
                     Assert-MockCalled Remove-CMBoundaryFromGroup -Exactly -Times 1 -Scope It
+                    Assert-MockCalled Get-CMSiteSystemServer -Exactly -Times 0 -Scope It
+                    Assert-MockCalled Set-CMBoundaryGroup -Exactly -Times 0 -Scope It
                 }
 
                 It 'Should call expected commands when removing a boundary group' {
@@ -351,6 +590,8 @@ try
                     Assert-MockCalled Get-BoundaryInfo -Exactly -Times 0 -Scope It
                     Assert-MockCalled Remove-CMBoundaryGroup -Exactly -Times 1 -Scope It
                     Assert-MockCalled Remove-CMBoundaryFromGroup -Exactly -Times 0 -Scope It
+                    Assert-MockCalled Get-CMSiteSystemServer -Exactly -Times 0 -Scope It
+                    Assert-MockCalled Set-CMBoundaryGroup -Exactly -Times 0 -Scope It
                 }
 
                 It 'Should throw and call expected commands for adding a new boundary that does not exist' {
@@ -368,13 +609,236 @@ try
                     Assert-MockCalled Get-BoundaryInfo -Exactly -Times 1 -Scope It
                     Assert-MockCalled Remove-CMBoundaryGroup -Exactly -Times 0 -Scope It
                     Assert-MockCalled Remove-CMBoundaryFromGroup -Exactly -Times 0 -Scope It
+                    Assert-MockCalled Get-CMSiteSystemServer -Exactly -Times 0 -Scope It
+                    Assert-MockCalled Set-CMBoundaryGroup -Exactly -Times 0 -Scope It
+                }
+
+                It 'Should call expected commands for adding and removing Site Systems' {
+                    Mock -CommandName Get-TargetResource -MockWith { $getReturn }
+                    Mock -CommandName Convert-BoundariesIPSubnets
+                    Mock -CommandName Get-BoundaryInfo
+
+                    Set-TargetResource @matchSiteSystems
+                    Assert-MockCalled Import-ConfigMgrPowerShellModule -Exactly -Times 1 -Scope It
+                    Assert-MockCalled Set-Location -Exactly -Times 2 -Scope It
+                    Assert-MockCalled Get-TargetResource -Exactly -Times 1 -Scope It
+                    Assert-MockCalled New-CMBoundaryGroup -Exactly  -Times 0 -Scope It
+                    Assert-MockCalled Add-CMBoundaryToGroup -Exactly -Times 0 -Scope It
+                    Assert-MockCalled Convert-BoundariesIPSubnets -Exactly -Times 0 -Scope It
+                    Assert-MockCalled Get-BoundaryInfo -Exactly -Times 0 -Scope It
+                    Assert-MockCalled Remove-CMBoundaryGroup -Exactly -Times 0 -Scope It
+                    Assert-MockCalled Remove-CMBoundaryFromGroup -Exactly -Times 0 -Scope It
+                    Assert-MockCalled Get-CMSiteSystemServer -Exactly -Times 1 -Scope It
+                    Assert-MockCalled Set-CMBoundaryGroup -Exactly -Times 2 -Scope It
+                }
+
+                It 'Should throw and call expected commands when SiteSystemsInclude and SiteSystemsExclude contain the same entry' {
+                    Mock -CommandName Get-TargetResource -MockWith { $getReturn }
+                    Mock -CommandName Convert-BoundariesIPSubnets
+                    Mock -CommandName Get-BoundaryInfo
+
+                    { Set-TargetResource @siteSystemsIncludeExclude } | Should -Throw -ExpectedMessage $siteInEx
+                    Assert-MockCalled Import-ConfigMgrPowerShellModule -Exactly -Times 1 -Scope It
+                    Assert-MockCalled Set-Location -Exactly -Times 2 -Scope It
+                    Assert-MockCalled Get-TargetResource -Exactly -Times 1 -Scope It
+                    Assert-MockCalled New-CMBoundaryGroup -Exactly  -Times 0 -Scope It
+                    Assert-MockCalled Add-CMBoundaryToGroup -Exactly -Times 0 -Scope It
+                    Assert-MockCalled Convert-BoundariesIPSubnets -Exactly -Times 0 -Scope It
+                    Assert-MockCalled Get-BoundaryInfo -Exactly -Times 0 -Scope It
+                    Assert-MockCalled Remove-CMBoundaryGroup -Exactly -Times 0 -Scope It
+                    Assert-MockCalled Remove-CMBoundaryFromGroup -Exactly -Times 0 -Scope It
+                    Assert-MockCalled Get-CMSiteSystemServer -Exactly -Times 0 -Scope It
+                    Assert-MockCalled Set-CMBoundaryGroup -Exactly -Times 0 -Scope It
+                }
+
+                It 'Should throw and call expected commands when Site System does not exist' {
+                    Mock -CommandName Get-TargetResource -MockWith { $getReturn }
+                    Mock -CommandName Convert-BoundariesIPSubnets
+                    Mock -CommandName Get-BoundaryInfo
+                    Mock -CommandName Get-CMSiteSystemServer -MockWith { $null }
+
+                    { Set-TargetResource @includeSiteSystems } | Should -Throw -ExpectedMessage 't'
+                    Assert-MockCalled Import-ConfigMgrPowerShellModule -Exactly -Times 1 -Scope It
+                    Assert-MockCalled Set-Location -Exactly -Times 2 -Scope It
+                    Assert-MockCalled Get-TargetResource -Exactly -Times 1 -Scope It
+                    Assert-MockCalled New-CMBoundaryGroup -Exactly  -Times 0 -Scope It
+                    Assert-MockCalled Add-CMBoundaryToGroup -Exactly -Times 0 -Scope It
+                    Assert-MockCalled Convert-BoundariesIPSubnets -Exactly -Times 0 -Scope It
+                    Assert-MockCalled Get-BoundaryInfo -Exactly -Times 0 -Scope It
+                    Assert-MockCalled Remove-CMBoundaryGroup -Exactly -Times 0 -Scope It
+                    Assert-MockCalled Remove-CMBoundaryFromGroup -Exactly -Times 0 -Scope It
+                    Assert-MockCalled Get-CMSiteSystemServer -Exactly -Times 1 -Scope It
+                    Assert-MockCalled Set-CMBoundaryGroup -Exactly -Times 0 -Scope It
                 }
             }
         }
 
-        Describe "$moduleResourceName\Test-TargetResource" {
-            Mock -CommandName Set-Location
-            Mock -CommandName Import-ConfigMgrPowerShellModule
+        Describe 'ConfigMgrCBDsc - DSC_CMBoundaryGroups\Test-TargetResource' -Tag 'Test' {
+            BeforeAll {
+                $mockOutputBoundarySubnet = @(
+                    (New-CimInstance -ClassName DSC_CMBoundaryGroupsBoundaries `
+                        -Namespace root/microsoft/Windows/DesiredStateConfiguration `
+                        -Property @{
+                            'Value' = '10.1.3.0'
+                            'Type'  = 'IPSubnet'
+                        } `
+                        -ClientOnly
+                    )
+                )
+
+                $mockInputBoundarySubnet = @(
+                    (New-CimInstance -ClassName DSC_CMBoundaryGroupsBoundaries `
+                        -Namespace root/microsoft/Windows/DesiredStateConfiguration `
+                        -Property @{
+                            'Value' = '10.1.3.1/24'
+                            'Type'  = 'IPSubnet'
+                        } `
+                        -ClientOnly
+                    )
+                )
+
+                $mockBoundaryMembers = @(
+                    (New-CimInstance -ClassName DSC_CMBoundaryGroupsBoundaries `
+                        -Namespace root/microsoft/Windows/DesiredStateConfiguration `
+                        -Property @{
+                            'Value' = '10.1.1.1-10.1.1.255'
+                            'Type'  = 'IPRange'
+                        } `
+                        -ClientOnly
+                    ),
+                    (New-CimInstance -ClassName DSC_CMCollectionQueryRules `
+                        -Namespace root/microsoft/Windows/DesiredStateConfiguration `
+                        -Property @{
+                            'Value' = '10.1.3.0'
+                            'Type'  = 'IPSubnet'
+                        } `
+                        -ClientOnly
+                    ),
+                    (New-CimInstance -ClassName DSC_CMCollectionQueryRules `
+                        -Namespace root/microsoft/Windows/DesiredStateConfiguration `
+                        -Property @{
+                            'Value' = 'First-Site'
+                            'Type'  = 'ADSite'
+                        } `
+                        -ClientOnly
+                    )
+                )
+
+                $mockInputBoundaryADSiteAdd = @(
+                    (New-CimInstance -ClassName DSC_CMBoundaryGroupsBoundaries `
+                        -Namespace root/microsoft/Windows/DesiredStateConfiguration `
+                        -Property @{
+                            'Value' = 'Second-Site'
+                            'Type'  = 'ADSite'
+                        } `
+                        -ClientOnly
+                    )
+                )
+
+                $mockInputBoundaryADSite = @(
+                    (New-CimInstance -ClassName DSC_CMBoundaryGroupsBoundaries `
+                        -Namespace root/microsoft/Windows/DesiredStateConfiguration `
+                        -Property @{
+                            'Value' = 'First-Site'
+                            'Type'  = 'ADSite'
+                        } `
+                        -ClientOnly
+                    )
+                )
+
+                $mockInputBoundaryRange = @(
+                    (New-CimInstance -ClassName DSC_CMBoundaryGroupsBoundaries `
+                        -Namespace root/microsoft/Windows/DesiredStateConfiguration `
+                        -Property @{
+                            'Value' = '10.1.1.1-10.1.1.255'
+                            'Type'  = 'IPRange'
+                        } `
+                        -ClientOnly
+                    )
+                )
+
+                $getReturnAbsent = @{
+                    Site          = 'Lab'
+                    BoundaryGroup = 'TestGroup'
+                    Boundaries    = $null
+                    SiteSystems   = $null
+                    Ensure        = 'Absent'
+                }
+
+                $getReturn = @{
+                    Site          = 'Lab'
+                    BoundaryGroup = 'TestGroup'
+                    Boundaries    = $mockBoundaryMembers
+                    SiteSystems   = @('DP01.contoso.com', 'DP02.contoso.com')
+                    Ensure        = 'Present'
+                }
+
+                $setTestInputAdd = @{
+                    SiteCode       = 'Lab'
+                    BoundaryGroup  = 'TestGroup'
+                    Boundaries     = $mockInputBoundarySubnet
+                    BoundaryAction = 'Add'
+                }
+
+                $setTestInputAddValue = @{
+                    SiteCode       = 'Lab'
+                    BoundaryGroup  = 'TestGroup'
+                    Boundaries     = $mockInputBoundaryADSiteAdd
+                    BoundaryAction = 'Add'
+                }
+
+                $setTestInputMatch = @{
+                    SiteCode       = 'Lab'
+                    BoundaryGroup  = 'TestGroup'
+                    Boundaries     = $mockInputBoundaryRange
+                    BoundaryAction = 'Match'
+                }
+
+                $setTestInputRemove = @{
+                    SiteCode       = 'Lab'
+                    BoundaryGroup  = 'TestGroup'
+                    Boundaries     = $mockInputBoundarySubnet
+                    BoundaryAction = 'Remove'
+                }
+
+                $setTestInputGroupOnlyAbsent = @{
+                    SiteCode      = 'Lab'
+                    BoundaryGroup = 'TestGroup'
+                    Ensure        = 'Absent'
+                }
+
+                $setTestInputGroupOnly = @{
+                    SiteCode      = 'Lab'
+                    BoundaryGroup = 'TestGroup'
+                    Ensure        = 'Present'
+                }
+
+                $getReturnNoBoundaries = @{
+                    SiteCode      = 'Lab'
+                    BoundaryGroup = 'TestGroup'
+                    Boundaries    = $null
+                    SiteSystems   = @('DP01.contoso.com', 'DP02.contoso.com')
+                    Ensure        = 'Present'
+                }
+
+                $matchSiteSystems = @{
+                    SiteCode             = 'Lab'
+                    BoundaryGroup        = 'TestGroup'
+                    SiteSystems          = 'DP01.Contoso.com','DP03.contoso.com'
+                    SiteSystemsToInclude = 'DP02.contoso.com'
+                    SiteSystemsToExclude = 'DP03.contoso.com'
+                }
+
+                $siteSystemsIncludeExclude = @{
+                    SiteCode             = 'Lab'
+                    BoundaryGroup        = 'TestGroup'
+                    SiteSystemsToInclude = 'DP02.contoso.com'
+                    SiteSystemsToExclude = 'DP02.contoso.com'
+                }
+
+                Mock -CommandName Import-ConfigMgrPowerShellModule
+                Mock -CommandName Set-Location
+            }
 
             Context 'When running Test-TargetResource' {
 
@@ -446,6 +910,20 @@ try
                     Mock -CommandName Convert-BoundariesIPSubnets
 
                     Test-TargetResource @setTestInputGroupOnlyAbsent | Should -Be $true
+                }
+
+                It 'Should return desired result false when SiteSystems does not match' {
+                    Mock -CommandName Get-TargetResource -Mockwith { $getReturn }
+                    Mock -CommandName Convert-BoundariesIPSubnets
+
+                    Test-TargetResource @matchSiteSystems | Should -Be $false
+                }
+
+                It 'Should return desired result false when SiteSystemsToInclude and SiteSystemsToExclude contain the same value' {
+                    Mock -CommandName Get-TargetResource -Mockwith { $getReturn }
+                    Mock -CommandName Convert-BoundariesIPSubnets
+
+                    Test-TargetResource @siteSystemsIncludeExclude | Should -Be $false
                 }
             }
         }